Fix NodeLatencyMonitor not starting (#6552)
The NodeLatencyMonitor was not starting (more precisely, the monitorLoop
method was never called). This is because NewNodeLatencyMonitor was
called *after* starting the informer factory. This means that the
informer created in NewNodeLatencyMonitor was never started, and that
WaitForNamedCacheSync was blocking indefinitely.

With this change, we ensure that NewNodeLatencyMonitor is called
*before* starting the informer factory. The Run method is still called
at the end of Agent "initialization", once the datapath has been
configured.

Note that it is not clear that WaitForNamedCacheSync is really needed
for the NodeLatencyMonitor. However, it is probably fine to use the same
"pattern" as for our other controllers.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>