Use appropriate docker driver in Github worfklows
For most workflows, we want to use the "docker" driver, while the
"docker-container" driver should be used when images need to be pushed
to the registry (so we can use registry-based caching).

When using the "docker-container" driver, images are built in a build
container, and there is no access to the docker image store (see
https://github.com/moby/buildkit/issues/2343). This means that when
building a new image, we cannot use "local" images as base images. This
is an issue because the base image may not exist yet in the directory,
and because we may want to use a modified base image which is not the
same as the one in the registry.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>