Upgrade Ubuntu to 24.04 (Noble) and Suricata to 7.0
Upgrade to ubuntu:24.04 in most places (except for a few build / test
images which will be updated later on). In particular ubuntu:24.04 is
now the base image for antrea/antrea-agent-ubuntu and
antrea/controller-ubuntu.

In order for our current version of OVS (2.17.7) to build correctly on
Ubuntu 24.04, some "recent" patches have to be applied manually.

We also upgrade Suricata to 7.0, as OISF does not publish packages for
Suricata 6.0 for Noble. Suricata 7.0 includes support for some new
protocols, including HTTP/2 and QUIC, which means we could start
supporting these protocols in L7NetworkPolicies.

After this change, the antrea/antrea-agent-ubuntu image is slightly
bigger (344MB -> 368MB uncompressed), and the size increase seems to be
driven by some systemd libraries.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>