Use same MTU as uplink for bridge port
In bridging mode (on Linux), when moving the physical adapter to the
bridge, we explictly set the MTU for the bridge port to the same value
as for the physical adapter. Without this change, the MTU may default to
a different (lower) value if some existing container ports have a lower
MTU value. For example, this occurs when first installing Antrea in
encap mode, then re-installing Antrea in noEncap mode with bridging mode
enabled.

We also do some minor documentation updates to indicate to users that
they should consider restarting existing workloads when updating the
Antrea datapath configuration.

Fixes #6456

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>