Improve L7Listener code in FlowExporter (#6576)
The L7Listener is responsible for receiving and processing Suricata
events. When the stop channel is closed, the listener should be closed
in order to stop accepting connections. This doesn't really matter when
this code is run as part of the antrea-agent, but it is useful for unit
testing.

We also improve the unit tests for the L7Listener. Sub-tests should not
share a common gomock controller. Instead of sleeps, we use polling when
waiting for conditions to be met.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>