Support --random-fully for iptables SNAT / MASQUERADE rules (#6602)
We expose 2 new configuration parameters for the Agent:
snatFullyRandomPorts and egress.snatFullyRandomPorts. When the first one
is set to true, the MASQUERADE iptables rules used to implement "local"
Node SNAT will use randomized source port mappings. When the second one
is set to true, the SNAT iptables rules used to implement Egress SNAT
will use randomized source port mappings. This is achieved with the
`--random-fully` iptables flag.

These new configuration parameters are only supported on Linux
Nodes. They require iptables >= 1.6.2, which is not a problem with the
standard antrea-agent container images. They also require Linux kernel
>= 3.13 and >= 3.14 respectively, but these are very old releases so we
can ignore this requirment, which is consistent with the K8s
implementation.

snatFullyRandomPorts is set to false by default (this may change in the
future). egress.snatFullyRandomPorts is set to null (empty) by default,
which means that unless the parameter is explicitly set, we will use the
top-level snatFullyRandomPorts value.

Fixes #6544

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>