Fix L7NP enable logging wrong packet
Current logs by Suricata when enableLogging is set, logs the
wrong packet of RST instead of the original TCP packet, and
remains an existing bug in Suricata.

This solution modifies the design to remove enableLogging in
L7, keep it in L4, and instead configure Suricata to log Packet
info for all alert events.

Fixes #6636.

Signed-off-by: Qiyue Yao <yaoq@vmware.com>