[FlowAggregator] Add templateRefreshTimeout configuration
Can be used to configure the template retransmission interval when using
the UDP protocol to export records to an IPFIX collector.

We also ensure that the default value is 600s, not 1800s, which is the
value specified in RFC 6728. We do the same thing for the Agent
Exporter, but this time by letting the go-ipfix library decide the value
(we set it to 0 in ExporterInput to request the default). This is
important because the templateRefreshTimeout config in the Agent's
exporter and the templateTTL in teh Aggregator's collector must "match"
(templateTTL = 3 x templateRefreshTimeout), and so it is better to let
go-ipfix pick the right values for both.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>