Add missing metav1.ObjectMeta to NetworkPolicyEvaluation API type
It seems that omitting the field was a mistake when defining the API.
New versions of the K8s code generators (client-gen specifically) will
not work with types missing this field.
See https://github.com/kubernetes/code-generator/commit/688a73a18a8c19ffd7979f5c9a6c03d118cc4d7f
This is preventing us from updating to a more recent version of the K8s
libraries and of the code generators.

Note that I believe that this API was modeled after the upstream
TokenReview API (which like this one only supports the "create" verb,
and returns a response immediately). The TokenReview type does include a
metav1.ObjectMeta field.

Unfortunately, for backwards-compatibility, we need to use Protobuf
field number 3 for the metadata field, while K8s always uses 1. This is
a bit surprising, but it should only affect the wire format, and nothing
else.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>