Change antctl make target (#6686)
Running "make antctl" is common when someone wants to build antctl for
the current machine. The current target builds 3 versions of the antctl
binary: one for Linux, one for Windows and one for Darwin (using the
architecture of the current machine). This seems unnecessary and
potentially confusing. We change the antctl target to be an alias of
antctl-release. Running "make antctl" will now produce a single binary
named "antctl", which can be run on the current machine.

For backwards-compatibility, we preserve the ability to run "make
antctl-linux", "make antctl-windows", and "make antctl-darwin". These 3
targets combined will produce the 3 antctl binaries described above.

Finally, we update the Github workflow in charge of validating the
antctl Go build. We now have one dedicated step for each supported OS /
arch combination.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>