Fix flaky TestSubscribe test in pkg/util/channel
I have observed the test failing in Github CI, and I was able to
reproduce locally by setting GOMAXPROCS=2.

When using the race detector (-race flag), it usually takes 25ms for the
condition function in assert.Eventually to run successfully on my
laptop. However, it sometimes spikes to 100+ms, and then the test fails,
as the timeout is set to 100ms.

My take is that there could be too much contention between the channel
logic and the goroutine performing the checks. The issue only occurs
when using the race detector.

To avoid the issue, we use a timeout of 500ms and we set the tick period
to 50ms. This is more than sufficient on my machine. If we see the
failure in CI again, we can always double these values.

We also fix the error log for the test, as the previous log would never
print meaningful values: the log arguments are evaluatd when
assert.Eventually is called, and not after the assertion fails.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>