Block on stopCh in Run method of controllerMonitor (#6708)
The Run method of controllers usually never returns until the stop
channel is closed. The controllerMonitor seems to be the exception, as
it returns after "initialization", despite the Run method being invoked
in a separate goroutine. This leads to a confusing message in the Antrea
Controller logs, informing us that AntreaControllerMonitor is shutting
down.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>