Always require the cluster UUID in the FlowAggregator (#6714)
Before this change, the cluster UUID was strictly required by the
S3Exporter and by the ClickHouseExporter, but somewhat optional for the
IPFIXExporter (if not available after a certain timeout, a random UUID
was generated and used to compute the IPFIX observation domain
ID). Furthermore, every exporter by responsible for calling
getClusterUUID independently, and that was the only reason for their
implementations to require access to the K8s client.

I think it makes more sense to make things more uniform, and require the
cluster UUID to be available regardless of which exporters are
enabled. We can retrieve the cluster UUID once during FlowAggrgator
initialization, then pass it along to all exporters which need it.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>