Fix error logs for FlowAggregator e2e tests (#6730)
When the IPFIX collector has not received the expected flow records, we
try to dump the actual list of records to facilitate
troubleshooting. However, the output is to large to be handled by the
testify assertions (and IMO it is too large to be useful for
troubleshooting). With this change, if the expected records are not
found in the collector output, we only print the 20 most recent records,
in a format that should be easier to leverage for troubleshooting.

Fixes #6478

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>