Run conformance tests for Arm images in this repo
This is a follow-up to #6486, which added support for building Antrea
Arm images in the antrea-io/antrea repo, using Github-hosted Arm
runners.

With this change, we will also be testing Arm images in this repo,
hence finishing to remove our dependency on a private reposiotry, with
self-hosted runners (vmware-tanzu/antrea-build-infra).

This change is very small, as we re-use the existing conformance.yml
workflow, and add a dispatch step to the build.yml and build_tag.yml
workflows (same repo dispatch, instead of dispatching to
vmware-tanzu/antrea-build-infra, as was the case previously). Using
Sonobuoy, we run the conformance test suite every time the main branch
is updated, as well as for every release. Unfortunately, we have to
re-build the image before running the tests, but thanks to caching it
only represents a small amount of time.

We also remove some deprecated maintainer documentation.

Fixes #6453

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>