Enable disabling TX checksum offload for Antrea host gateway
This commit introduces the ability to disable TX checksum offload
for the host gateway interface (default: `antrea-gw0`) by setting the
`disableTXChecksumOffload` option to `true`.

Note: If this option is later set to false, Antrea will not restore
the original TX checksum state, as it does not retain the original
configuration. Users are responsible for manually reconfiguring the
setting if needed.

Signed-off-by: Hongliang Liu <hongliang.liu@broadcom.com>