Push Antrea Ubuntu-based images to ghcr.io (#6834)
The Github Container Registry (ghcr.io) is introduced as an alternative
to Docker Hub for users. There is no change to how images are pushed to
Docker Hub, but antrea/antrea-agent-ubuntu and
antrea/antrea-controller-ubuntu are now also available in ghcr.io, as
antrea-io/antrea-agent-ubuntu and antrea-io/antrea-controller-ubuntu
respectively.

We do not retroactively push previously-released images to
ghcr.io. Antrea v2.3 will be the first release also available in ghcr.io.

In the future, we may also push UBI-based images to ghcr.io.

Fixes #6454

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>