Add Jenkins job for scaling agents
* The script facilitates scaling of Jenkins agents by allowing both adding agents and removing agents.
It integrates with GCP and the Jenkins Swarm Plugin to add or remove agents dynamically based on the queue load, improving testbed efficiency.

* The `add_agent` function scales up the Jenkins agents by checking if jobs with the specified label are present in the Jenkins queue.
If jobs are found, it creates a new GCP VM, waits for the external IP to become available, and registers the new agent with Jenkins.

* The `remove_agent` function scales down the Jenkins agents by identifying idle agents.
It removes the agent from Jenkins and deletes the associated VM, ensuring that only idle agents are removed, preventing any disruption to ongoing jobs.

Signed-off-by: Shuyang Xin <shuyang.xin@broadcom.com>