[Windows] Support OF port state "down" when installing Pod forwarding rules
On Windows, OVS has an issue which doesn't correctly update the port status
after an OpenFlow port is successfully installed. So OVS may send out PortStatus
message with Port state as LIND_DOWN, this issue doesn't impact on the datapath
forwarding.

This change is a workaround to ensure Pod's OpenFlow entries are installed as
long as the OpenFlow port is allocated.

Signed-off-by: Wenying Dong <wenyingd@vmware.com>