Ensure that promote_secondaries is set on IPAssigner interfaces (#6898)
The IPAssigner should ensure that the promote_secondaries sysctl
variable is always set when creating an interface. This ensures that
When the primary IP address is removed from the interface, a secondary
IP address is promoted, instead of removing all the corresponding
secondary IP addresses. Otherwise, the deletion of one IP address can
trigger the automatic removal of all other IP addresses in the same
subnet, if the deleted IP happens to be the primary (first one assigned
chronologically). For example, this can affect the Egress feature (when
EgressSeparateSubnet is used).

Fixes #6890

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>