Remove local ASN range limitation in BGPPolicy (#6914)
Previously, for safety and internal using only purposes, the local ASN was
restricted to the private range (64512-65535). However, based on user feedback,
BGPPolicy may also be used to peer with public BGP ASNs. Therefore, this
restriction is no longer justified.

To prevent potential misconfigurations that could disrupt BGP operations,
we have added the follow paragraph to documentation to remind users of the
implications of using unrestricted ASN ranges.

```
Private ASNs, which are within the ranges 64512-65534 (16-bit), should be strictly limited to private networks or
environments that do not peer with public ASNs. If public network connectivity is required, coordinate with your upstream
provider to avoid issues caused by private ASN usage.
```

Signed-off-by: Hongliang Liu <hongliang.liu@broadcom.com>