Add antreaProxy.disableServiceHealthCheckServer config (#6939)
Running both kube-proxy and Antrea Proxy with `proxyAll` can trigger
some error logs for Services of type LoadBalancer with
`externalTrafficPolicy` set to `Local`. For such Services, the proxy is
in charge of running a health check server on each Node, in order to
report the number of local Endpoints which implement the Service.
Because both kube-proxy and Antrea Proxy (with `proxyAll`) will try to
run the health check servers on the same ports, one of them will fail to
bind to the desired address. In practice, we typically observe that
Antrea Proxy tries to bind to the address first (and succeeds), while
kube-proxy fails and logs an error message, but there is no guarantee
that it will be the case.

To avoid kube-proxy error logs, users can now set
`antreaProxy.disableServiceHealthCheckServer` to true in the Antrea
Agent's config, which will instruct Antrea Proxy to stop running health
check servers for LoadBalancer Services. This is not a perfect solution
as ideally the component responsible for the proxy implementation
(Antrea Proxy in this case) should also be responsible for providing
health check information.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>