Add Proxy mode for FlowAggregator (#6920)
In Proxy mode, connection data is not buffered in the FlowAggregator and
records from source and destination Nodes are not correlated. Instead,
records are proxied directly to the destination collector. In Proxy
mode, only the IPFIX exporter is supported.

A simple e2e test is added to test this new mode.

Fixes #6773

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>