Handle Pod UID updates in PodStore
The PodStore uses the UID as the unique key in the indexer, and not the
usual namespace/name. Because it is possible for the deletion of a Pod,
followed by the creation of a different Pod wih the same name, to be
observed as an Update by a client (see
https://pkg.go.dev/k8s.io/client-go/tools/cache#SharedInformer), the
Update event handler needs to handle this case gracefully by comparing
UIDs. Note that this is probably quite an unlikely situation in the case
of Pods, but we should still account for this in the PodStore
implementation.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>