Add checksum/config annotation to FlowAggregator
When updating the config with Helm, the annotation will be automatically
updated and the FlowAggregator Pod will be recreated. This is consistent
with the main Helm chart for Antrea.

Note that the FlowAggregator can handle some config changes
"dynamically" (i.e., without a Pod restart), but not all config updates
are supported and we may want to remove that partial support altogether
in the future.

The annotation can also be updated manually to roll the Deployment (e.g,
if Helm is not being used).

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>