Automated cherry pick of #6821: Fix packetcapture bpf filter issue (#6815) (#6821)#6804: Fix PacketCapture pcapng file issue on macOS (#6804) (#6983)
* Fix packetcapture bpf filter issue (#6815) (#6821)

In PacketCapture, packets which don’t match the target BPF can be
received after the socket is created and before the bpf filter is
applied. This patch uses a zero bpf filter (matches no packet), then
empties out any packets that arrived before the "zero-BPF" filter was
applied. At this point the socket is definitely empty and it can’t
fill up with junk because the zero-BPF is in place. Then we replace
the zero-BPF with the real BPF we want.

Signed-off-by: Hang Yan <yhang@vmware.com>
Co-authored-by: Antonin Bas <antonin.bas@broadcom.com>

* Fix PacketCapture pcapng file issue on macOS (#6804)

By default, gopacket will write snap length=0 in the pcapng file
header, means unlimited snaplen. tcpdump on macOS (libpcap version 1.10.1)
cannot recognize this and will report error. This patch sets
a default value (65535) for it.

For #6795

Signed-off-by: Hang Yan <yhang@vmware.com>

---------

Signed-off-by: Hang Yan <yhang@vmware.com>
Co-authored-by: Antonin Bas <antonin.bas@broadcom.com>