[FlowAggregator] Use correct input channel in flowExportLoopProxy
The code was invalid as the function was consuming messages straight
from the collector instead of from the preprocessor, hence bypasing the
preprocessor. This was not caught by e2e tests because the preprocessor
output channel is small (capacity of 16), hence it was filling up very
quickly and only a few uninteresting (i.e., irrelevant to the test
cases) records were lost.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>