[FlowAggregator] Accept configurations with no collector / sink
Prior to this change, the FlowAggregator would crash if the
configuration didn't enable at least one collector (e.g., the IPFIX
collector). While it was probably introduced to help identify
misconfigurations easily, it seems like an unnecessary restriction. We
should be able to deploy the FlowAggregator in such a way that IPFIX
records are still received from Agents (and metrics are incremented),
while discarding aggregated / proxied data instead of exporting it.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>