Fix glob pattern in codecov action (Kind tests)
The codecov CLI has been updated, and the glob translation logic is
different. We need the '**' pattern to match the full path.
See https://github.com/codecov/codecov-cli/pull/619

We also remove the trailing wildcard, as it seems that we only ever use
the .cov.out suffix for these tests.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>