Compare PodCIDRs in NodeRouteController
When determining whether a Node's configuration has changed.

Note that when writing a controller, we should never assume that when
the object name is the same, all immutable fields are guaranteed to be
the same, and as a result we should always check all of them. This
applies even for controllers without a workqueue (which is not the case
here), as a Delete followed by a Create with the same name can show as
an Update event. It is more likely to happen when the controller uses a
workqueue, as in that case, even if the Delete and Add event handlers
are called sequentially, the key may only be processed once, after the
informer store has been updated with the new object.

Fixes #6965

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>