Add more configuration flexibility to Egress for user
Added a StrictEnforcement field to the Egress spec.
If this field is set to flase then if the egress node
is not available due to any reason then the packet
transfer or we can say the traffic will go via normal
Node SNAT, while in the second case where this field
is set to true, and if egress node is not available
packet won't go out and traffic will be stuck
because of unavailability of egress node.

Signed-off-by: Pulkit Jain <pulkit.jain@broadcom.com>